Changes necessary to implement the "horizontal workflow" portion of Spectroscope. Within each cluster, the latency distribution of each edge in the problem period is compared with that of the non-problem period by using a simple statistical test. Edges whose distributions have changed are high-lighted in red.
In addition to the above, the code in this commit overlays per-cluster aggregate
information about each edge on top of the of the cluster visualization.
This information includes:
   p: probability of a false-positive,
   a: average latency in non-problem period/average latency in problem-period
   s: standard deviation in non-problem period/standard-deviation in problem period

Finally, boxplots of request response-times are automatically generated for each
cluster and stored as PNG files in results/boxplots.


---
Note: The edge comparison code still doesn't work perfectly -- it's currently
way too aggressive in concluding that two edge distributions are
statistically different.  I'm going to try to figure this out today, but I wanted
to commit the basic structure of the code, as the structure/basic code quality
is pretty sound.