The goal of highlighting edges in red is to identify the most conspicuous offenders (the edges that we know for a fact have changed for the worse in terms of their latency), not all of them.  So, instead of comparing raw values in microseconds, compare a smoothed version (floor time_in_us/1000).  At least with the RMW problem, this keeps the test from being overly aggressive.
Not sure if the above heuristic will work in general, but it's not worth trying
all the alternatives yet.  I'd rather move on to the decision tree and get to
actual problems quickly.

An interesting question is why not simply compare averages instead of using a
statistical test.  The answer is that a statistical test should optimally trade
off confidence in the result with the variance and amount of data presented to
it.  So, basically, if you see an edge highlighted in red you should know for a
fact that it's something you need to look into.  Of course, with the amount of
finessing required for the statistical test, you gotta wonder if just visually
looking at the average might work out better.

Along w/implementing the decision tree, I'm going to try and get the existing
Spectroscope code into the nightly runs (to analyze sfs and postmark) asap.
This is probably the easiest way to figure out of the workflow/algorithms are
useful.