* Fix in PrintRequests.pm: When a semantic label is not present, parse_nodes() incorrectly tried to use the non-existent semantic label as the node name.  This may have had something to do with the fact that every now and then, certain edges were not overlayed w/the aggregate edge information.
* When extracting edge latencies for a request, the function in PrintRequests.pm would
use $_ instead of the global_id of the request when obtaining the string representation
of the request.  $_ was not guaranteed to hold the right glboal ID.  This might have
had something to do w/the non-smoothed version of the edge latency data leading KSTest
to always declare two latency distributions as different.  (this might or might be true).

* Part of the decision tree data extraction code is also present in this commit.
Specifically, the code to compare two graphs and generate a "template" list of
nodes that match between the two has been added.  I haven't yet committed the
code that uses this template to traverse through the cluster graphs and extract
the low-level data.

Moved some common functions into DotHelper.pm.