* Move graph matching code from PrintRequests.pm to MatchGraphs.pm
* Add "build_structured_request_given_global_id" call to PrintRequests; this
returns a tree representation of a DOT graph.

* Add working portion of infrastructure/code for regression tree classification
    * explain_clusters.pl is top-level script.

* Change ParseClustering results so that cluster representatives are
  printed out in "Cluster ID" order, instead of in some ranked order.  This
makes it easier for the user to identify the IDs of the candidate
mutated/original clusters when calling the explain_clusters.pl script.

* Change parse_nodes_from_string to use a concatenated string
"breadcrumb.timestamp" as the key when constructing the hash of node names.
Before, the key was an integer made up of the concatenation of
breadcrumb.timestamp.  The string representation preserves the distinction
between the breadcrumb and timestamp and so allows one to pull out the
breadcrumb at any point.


---
Note that APIs for lot of the perl modules could use some revision at this point, but will come back to this later.