This commit does not fix any correctness problems; it's just a change that should be made.
The "Alphabet Counter" used to be used to map nodes in the graphs to
characters; now it is used to simply map them to integers.  This counter used to start at "33" because this corresponds to the first printable character.
However, there is no reason to start at 33 anymore, since the mapping used
now is just integers.


In the future, it might be a good idea to rename the functions that create and
print the node mapping from "create_alphabet_mapping_*" and
print_alphabet_mapping_*" to something more generic.