There are two major bugfixes in this commit:
1)Implemented "Hack fix" for Self-* Bug 604 (Some edges are not overlayed
w/aggregate information) by reversing the order of the two nested for loops in
_overlay_edge_info().  Aggregate information about each edge (average latencies,
standard deviations, and the result of the hypothesis test) is computed on a
per-cluster/per unique "edge_name" basis.  This means that if the same edge name
appears in multiple locations within a single request (e.g., in a loop), this
code will compute the "aggregate" edge info by combining all the occurences of
this edge name.  This combined "aggregated" info is then overlayed on each
instance of the edge.  A similar case applies if an edge seen in different
locations across requests.

In effect, this implementatoin is "location agnostic."  It will tell you if a
particular interaction has changed in cost, but if there are multiple instances
of this interaction that occur at different points in a request, it will not be
able to tell you which instance changed.  I think This is the right "simple"
implementation for now, it's just I never thought of the implications consciously
before :)  In the future, we can move to a location-aware
implementation if need be.

2)Fixed a bug in which edges that spanned components would retain their original
edge label and would not be overlayed w/aggregate info in the Spectroscope
output.  This is almost moot because edges that span components are always marked
with an edge latency of zero, since we are not assuming synchronized times
across machines.  However, w/o this fix, Spectroscope's output visualization
"looked" wrong.

For each cluster, two sparse matrix files are created for input into the
hypothesis test in MATLAB.  Each line of the Sparse matrix files is of the
following format: "row num, col num edge latency".  Each row corresponds to a
unique edge and each column, to an occurence of that edge.  Matlab when
automatically convert a sparse row into a vector of zeros when you try to
access the row.

To maximize "debugability," the code in ParseClusteringResults.pm was written
so that the row numbers assigned to an edge would be the same across all clusters;
this way the sparse matrix created for different clusters would be comparable
(even though there is no reason they need to be).  This meant that the hypothesis
test matlab script has to "delete" materialized sparse rows internally.  As a result,
edges that span components (which always have a latency of zero) were indistinguishable
from sparse rows, were "deleted," and never showed up in the edge_list_hash used to
overlay the aggregate information.

The fix for this bug was to remove the global row number requirement (this
involved percolating API changes through several of the ParseClusteringResults
and PrintResults) functions.


Other changes:
--------------

* Installed "define.pm" in my perl installation, which adds
functionality equivalent to C-Preprocessor statements (e.g., #ifdef DEBUG)

* Added global constant DEBUG and print statements in the various modules
that will trigger if DEBUG => 1.