Last fix: my previous commit assumed that we could skip remvoing "0" values in the columns of the materialized sparse matrix.  Removing these "0" values is what caused the inter-component edges to be deleted, as their rows were always completely 0.
we can't skip removing "0" values, as matlab adds extra trailing zeros to each
row when materializing it to make sure the number of columns are the same across
all the rows of the sparse matrix.  Since the number of columns in each row
corresponds to the number of times an edge latency was observed for a given edge,
these zeros artificially deflate the mean.

The new solution is to delete trailing 0 values in each row and simply print
out an artificial result if the problem dataset and non-problem dataset
rows turn out to be empty.  The artifical result keeps the row from being 'deleted.'