Also need to exclude the "optype."  Since the optype is actually represented in the name of the nodes, two requests in different clusters will be perfectly seperated by this attribute.
The decision tree implementation now works for the RMW problem and separates
requests in the mutated cluster and original cluster based on the "count"
attribute of the top-level node.  Specifically, if count <=2048 -> problem, else
if count > 2048 -> no problem.

Admittedly, the RMW problem is a very simple test of the decision tree, as the
attributes that separates the clusters shows up in the root node.  For more
complex problems, I wouldn't be surprised if the tree has to be used in
iterative fashion.  That is, the programmer would run it once, check the
genrated rules. If the splits generated don't help him solve hte problem, he
should remove the root-node of the tree from the attribute list and then run the
algorithm again, etc.

I won't do this today, but I also need to write a script that, given a column
name, will remove that column from the decision tree input data table.  Right
now, you have to run the entire codebase again, meaning that all the SQL queries
need to be re-issued.