Change Spectroscope wrapper script and supporting libraries to support multiple problem and non-problem periods.  I'm a bit concerned that there might a decent amount of structural and time-related variance between independent runs of the same benchmark.  We'll have to investigate how much variance there is at some point.
Start working on "revision 2" of the codebase.  "Revision 1" is now a separate
branch (rajas-r1) and pretty much includes the functionality implemented for
the visit day demo (+some bug fixes).

Here's what I'm hoping for in this new version
   1) Support multiple problem/non-problem periods (duh)

   2) Automatically identify latency-based mutations

   3) Automatically supress output about "non-interesting" types
      * idea 1: Ignore request types that have even probability of
                occuring in either period, as long as they aren't
                latency-based mutations.

                If            P(request_type | problem_period)
                a_constant >  ----------------------------------   < b_constant
                            P(request_type | non_problem_period)

           then supress it, as long as it is not a latency-baed mutation
           [other more sophisticated metrics possible, need to think about it a bit]

   4) For each structural mutation, identify all types prevalent
     in the non-problem period  w/the same high-level request-type.

   5) Use a distance metric to rank these candidate originating clusters

   6) Install visualization for comparing the structural mutation
     to the candidate originating clusters.

   7) Install visualization for looking at latency-based mutations


All this is so that we can look at problems that manifest in workloads that
generate a lot of unique low-level request-types.  For example, just linux-build seems
to generate about 200 different types.  There might be an argument for clustering before
step 2, but I'd rather try the simpler approach first.

It'll also be useful to break the "200 types" into reads/writes, etc. and for
each of these high-level types, figure out how different their constituent
request-flow graphs tend to be.  What is the distribution of writes or reads by
request-flow type?  These answers would help guide the above heuristics.