Fixed a lot of small bugs.  Code now runs through to completion w/o crashing and generates results!  Still have to look at the results to see if they make sense :)
Internal notes:
---------------

The current method for identifying structural mutations and originating clusters
will balk when only one dataset is used as the input to Spectroscope.  This is
because P(cluster|problem period) is undefined in this case.  I could always
assign this probability a small non-zero value, but 'structural mutations' and
'originating clusters' have no meaning when only one dataset is present anyway.
Should just add a way to short-circuit indentification of strucutral
mutations/originating clusters when there's only one dataset.

Currently three output files are generated:
   * Graphs of structural mutations
   * Graphs of originating clusters
   * Graphs of clusters whose response-times have changed in distribution

Should add another output file that just lists all the clusters and info about them.

Need to modify the "summary node" added to each graph in the output files.  The
summary node should identify whether or not the cluster represents a change in
response-time, a structural mutation, or a originating cluster.  It should also
list the candidate originating clusters.  This information is obviously already
computed, I just need to change the summary node function to extract it.

Not sure if I should filter out 'noise': clusters that are technically
structural mutations or originating clusters, but which contain very very few
requests?  Obviously, want to keep these if their response-times are abnormally
high...