Finally, this is the initial commit for Spectroscope Version 3.
1)Infrastructure for plugging in different policies/heuristics for finding
  originators of a structural mutation.

2)Heuristic for determining the overall error when picking a particular
probabilistic mapping of mutations to originators.

3)Ability to rank structural mutations and response-time changes together.  That
is, the contribution of a response-time change to the performance change can now
be compared to that of a structural mutation.

There are weird 'edge cases.'  For example, what if a request-flow is marked as
both an originator *and* as a response-time change?
 In such cases, I assume that we can superimpose results.  When calculating The
cost of structural mutations that map to this originator, I use the originator's
response-time in the problem period. The cost of the response-time change is the
change in response-time between the problem and non-problem * the number of
requests from the non-problem period.  This means that the performance benefit
the system expects from the developer fixing the response-time change will only
be realized if hte corresponding structural mutations are fixed.  I need to
think more about such cases.


4)Very simple policies for picking originators (uniform, weighted by graph
closeness, etc.)  Goal is to see how poorly (or well) these very simple policies
do before moving to something more complex.

5)Modified version of refactor_graphs script, which given a set of refactorable
components, compresses all activity in insantiations of them into one edge.  In
addtion to bug fixes, this script now keeps the root node name in the refactored
version of the graphs.

6)New script convert_to_subgraphs.pl.  DOT seems to work much better when we
visualize request-flow graphs as 'one large graph with many disconnected
subgraphs.'  Changes to StructuredGraph.pm to support this change.  Basically,
when StructuredGraph's routines are called to create the refactored graphs, it
must guarantee that the node ids assigned each new graph are unique, or else
they will be connected in teh subgraph representation.