Adding these X^2 tests to my 'sandbox' directory.  Basically, the tool needs a good test that tells it whether or not structural mutations exist.
So far, I've been using a pretty standard X^2 test.  Since this test
determines whether two categorial distributions are the same, it essentially
determines whether structural mutations exist or not.  It does not yield
information about which categories are structural mutations or originators.

The test can be biased by a small number of categories for which the bin counts
are very different (the x^2 statistic is large).  So, if there are one or two
categories that differ a lot in their counts, the X^2 test might determine the
distributions are different, even if the other categories differ only very
slightly.

Usually, the above would not be a problem.  But, for some reason (yet to be
determined), the paths CREATEs and READDIRPLUSs take through the system seem to
vary enough to bias the X^2 test when comparing two executions of the same
benchmark.  This is when no performance problem is induced.

I need to figure out where these paths differ.  I can somewhat believe the
READDIRPLUS case -- in some sense, it's multiple requests batched in one.  But,
I need to figure out what's going on for the creates.

More significantly, I need to figure out what to do about the hypothesis test.
There are two options:
   1)Simply combine all readdirplus and creates into two unique categories before
     performing the X^2 test, or ignore them all together.  This has the effect of
     ignoring the changes in the distribution of hte way these requests are serviced.

   2)Create a variant of the X^2 test that determines whether individual categories
     have changed significantly in proportion.  This sort of test would try to identify
     which categories are mutations/originators directly, and would allow us to
     evaluate the test in terms of false positives, etc.