Implement inner loop of distance calculation in C; hook this into perl code via use of pipes.  This improves speed by over an order of magnitude, as the algorithm involves nested for loops, which perl is bad at because it's an interpreted language.
Edit distance calculation is still slower than I'd like though, mostly because
we calculate all (n^2)/2 possible distances (where N is the number of
categories).  In actuality, not all of them are needed --- we only need to
compute them between structural mutations and their possible originators.
Hooking this 'fast path' in and making it the default option during runtinme is
something I intend to do in the next commit.  Of course, sometimes its useful to
compute all of the edit distances, so the current way of doing things will
remain an option.  If we find in the future that we always want to compute edit
distances between every single category, we might consider implementing a
long-lived database to store these values across runs.

This is part 1 of a 3 part plan to improve the speed of the spectroscope
codebase, so that it will be easier to use it for diagnosing problems this time
around.

   * Part 2: Change code so that, by default, hypothesis tests for identifying
     edge latency changes are run only for categories that are identified as
     response-time mutations, and not for all categories.

   * Part 3: Change code so that response-time mutation and edge-latency
     mutation identification occurs  within a single matlab run; currently,
     matlab is started up (2*number of categories) times.

The goal is to finish these by the weekend, and to start evaluating
the tool on existing (and new) performance problems by Monday.