Seemingly working modifications necessary to run *ALL* hypothesis tests within one call to MATLAB.  Previously, MATLAB was called twice per category.  Since matlab takes ~3-5 seconds to start up and there can be many hundreds of categories, this was a huge overhead.  For example, SFS generates close to a 1000 categories, making the minimium Matlab overhead ~50 minutes(!).  However, calling Matlab once per category allowed for much easier debugging :)
Still needs some testing and modifications.  Most notably, I'm not yet sure what
to do about the X^2 test for testing the distribution of category counts between
snapshots.  It never worked, so I'm debating I should take it out completely, or
hack it into the new framework in hopes that some permutation of it will be
useful in the future.  The latter is not a large modification.

Detailed notes:

HypothesisTest.pm: A hypothesis test object now supports "multiple comparisons",
where each comparison encapsulates testing a group of distributions for
differences.  This is the abstraction that allows Matlab to be only called once
for all of statistical tests of the same type.  Basically, all hypothesis tests
for a category (response-time and edge latencies) map to one "comparison."
Multiple comparisons can be added before calling the "run_kstest2" function,
which actually initiates the hypothesis tests.

CreateHypothesisTests.pm: Has been modified to serve as a helper file for
ParseClusteringRresults.pm.  It supports a "add_latency_comparison()" function,
which does the job of formatting the per-category input files for the
HypothesisTest object and calling
"hypothesis_test->add_comparison(null_distribution_file,
test_distribution_file)."  For edge latencies, the mapping of edge names to row
numbers in the input files is now *global* across all categories and is passed
into "add_latency_comparison()" from the caller.  This means the input files
created for some categories will not be contiguous in their row numbering.  This
works, because the input file is specified in Matlab sparse matrix format; it is
the responsibility of the underlying matlab scripts to detect non-existent rows
and not consider them.

* ParseClusteringResults.pm: This module's "compute_cluster_info()" function has
been changed so that functions necessary to run the hypothesis tests for
identifying response-time mutations and edge latency changes are called in a
seperate second loop through the cluster_info hash.  In the first loop,
supporting information is collected, and the data files necessary to run the
hypothesis test are created via calls to
CreatHypothesisTests::add_latency_comparison();

* compare_edges.pm: Very slight changes to support modified API.  To support the
global mapping of edge names to rows, it expressly detects empty rows and skips
them completely.

* run_hypothesis_tests.m: This is a "meta file" that lists the input files and
output files for each comparison, and the name of the test that is to be run.
For each "comparison," it calls compare_edges.m with the appropriate
information.