Modify refactor_path.pl's logic to properly deal with concurrency and joins.
* Move "case 2" (original graph's destination node and refactored graph's
current outstanding node both refer to the same component) and "case 3"(original
graph's destination node and refactored graph's current outstanding node refer
to different components) to seperate functions.

* Get rid of 'global' mapping of original_node_id to refactored_node_id; this hash
allows the code to detect previously traversed paths (as will occur after a join) and
do the right thing.  To keep memory requirements low, this should be a per-request hash.

* If we see a previously traversed path, no need to go down it again(!)

* Deal with tough cases of Joins and Parallelism within components, instead of
just at component boundaries.  To deal with such cases, the code needs to look
at the number of parents and children the original graph's current destination
node posesses.  If either or both are greater than one, the refactored graph
needs a 'placeholder' 'split' or 'join' node or 'splitjoin' node.  This is true
even if the component is to be refactored.  Otherwise, latencies cannot be
accumulated properly.